# api/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Library management system with clean architecture",
        "title": "Library Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@library.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/authors": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of all authors in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authors"
                ],
                "summary": "List all authors",
                "responses": {
                    "200": {
                        "description": "List of authors",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/library-service_internal_adapters_http_dto.AuthorResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/library-service_internal_adapters_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/books/{id}/authors": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "books"
                ],
                "summary": "List authors of a book",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Book ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/library-service_internal_adapters_http_dto.AuthorResponse"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/library-service_internal_adapters_http_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/library-service_internal_adapters_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/payments/callback": {
            "post": {
                "description": "Process callback from payment gateway",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Handle payment callback",
                "parameters": [
                    {
                        "description": "Payment callback data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/library-service_internal_adapters_http_dto.PaymentCallbackRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/library-service_internal_adapters_http_dto.PaymentCallbackResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/library-service_internal_adapters_http_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/library-service_internal_adapters_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/payments/member/{memberId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all payments for a specific member",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "List member payments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Member ID",
                        "name": "memberId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/library-service_internal_adapters_http_dto.ListPaymentsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/library-service_internal_adapters_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/payments/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the current status of a payment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "payments"
                ],
                "summary": "Verify payment status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Payment ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/library-service_internal_adapters_http_dto.PaymentResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/library-service_internal_adapters_http_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/library-service_internal_adapters_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "library-service_internal_adapters_http_dto.AuthorResponse": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "pseudonym": {
                    "type": "string"
                },
                "specialty": {
                    "type": "string"
                }
            }
        },
...
        "library-service_internal_adapters_http_dto.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "library-service_internal_adapters_http_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/library-service_internal_adapters_http_dto.ErrorDetail"
                }
            }
        },
...
        "library-service_internal_adapters_http_dto.ListPaymentsResponse": {
            "type": "object",
            "properties": {
                "payments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/library-service_internal_adapters_http_dto.PaymentSummaryResponse"
                    }
                }
            }
        },
...
        "library-service_internal_adapters_http_dto.PaymentCallbackRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Transaction amount",
                    "type": "integer"
                },
                "approvalCode": {
                    "description": "Approval code",
                    "type": "string"
                },
                "cardMask": {
                    "description": "Masked card number",
                    "type": "string"
                },
                "code": {
                    "description": "\"ok\" or \"error\"",
                    "type": "string"
                },
                "currency": {
                    "description": "Currency code",
                    "type": "string"
                },
                "extra": {
                    "description": "Additional fields",
                    "type": "object",
                    "additionalProperties": true
                },
                "invoiceId": {
                    "description": "Order number",
                    "type": "string"
                },
                "reason": {
                    "description": "\"success\" or error description",
                    "type": "string"
                },
                "reasonCode": {
                    "description": "Error numeric code",
                    "type": "string"
                },
                "reference": {
                    "description": "Reference number",
                    "type": "string"
                },
                "terminal": {
                    "description": "Terminal ID",
                    "type": "string"
                },
                "transactionId": {
                    "description": "Gateway transaction ID",
                    "type": "string"
                }
            }
        },
        "library-service_internal_adapters_http_dto.PaymentCallbackResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "payment_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/library-service_internal_domain_payment.Status"
                }
            }
        },
        "library-service_internal_adapters_http_dto.PaymentResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "approval_code": {
                    "type": "string"
                },
                "card_mask": {
                    "type": "string"
                },
                "completed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "error_code": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "gateway_transaction_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invoice_id": {
                    "type": "string"
                },
                "member_id": {
                    "type": "string"
                },
                "payment_method": {
                    "$ref": "#/definitions/library-service_internal_domain_payment.PaymentMethod"
                },
                "payment_type": {
                    "$ref": "#/definitions/library-service_internal_domain_payment.PaymentType"
                },
                "related_entity_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/library-service_internal_domain_payment.Status"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "library-service_internal_adapters_http_dto.PaymentSummaryResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "completed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invoice_id": {
                    "type": "string"
                },
                "payment_type": {
                    "$ref": "#/definitions/library-service_internal_domain_payment.PaymentType"
                },
                "status": {
                    "$ref": "#/definitions/library-service_internal_domain_payment.Status"
                }
            }
        },
...
        "library-service_internal_domain_payment.PaymentMethod": {
            "type": "string",
            "enum": [
                "card",
                "wallet"
            ],
            "x-enum-varnames": [
                "PaymentMethodCard",
                "PaymentMethodWallet"
            ]
        },
        "library-service_internal_domain_payment.PaymentType": {
            "type": "string",
            "enum": [
                "fine",
                "subscription",
                "deposit"
            ],
            "x-enum-varnames": [
                "PaymentTypeFine",
                "PaymentTypeSubscription",
                "PaymentTypeDeposit"
            ]
        },
...
        "library-service_internal_domain_payment.Status": {
            "type": "string",
            "enum": [
                "pending",
                "processing",
                "completed",
                "failed",
                "cancelled",
                "refunded"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusProcessing",
                "StatusCompleted",
                "StatusFailed",
                "StatusCancelled",
                "StatusRefunded"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}